separate density estimation stage from torch model
